CS : challenge et register

+++ client/SocketHandler.cpp
@@ -47,8 +47,8 @@ void SocketHandler::set_session(Session* session)
 
 bool SocketHandler::recv_soft(char* buf, size_t len)
 {
-	int bytesRcvd = 0;
-	int bytesLeft = len;
+	size_t bytesRcvd = 0;
+	size_t bytesLeft = len;
 	int nBytes;
 
 	while (bytesRcvd < len)
@@ -67,8 +67,8 @@ bool SocketHandler::recv_soft(char* buf, size_t len)
 
 bool SocketHandler::send_soft(char* buf, size_t len)
 {
-	int bytesSent = 0;
-	int bytesLeft = len;
+	size_t bytesSent = 0;
+	size_t bytesLeft = len;
 	int nBytes;
 
 	while (bytesSent < len)
